Move tracking of builds to unit-test solution crawler code
