+++ src/EditorFeatures/Core/Remote/SolutionChecksumUpdater.cs
@@ -25,12 +25,6 @@ internal sealed class SolutionChecksumUpdater
 {
     private readonly Workspace _workspace;
 
-    /// <summary>
-    /// We're not at a layer where we are guaranteed to have an IGlobalOperationNotificationService.  So allow for
-    /// it being null.
-    /// </summary>
-    private readonly IGlobalOperationNotificationService? _globalOperationService;
-
     private readonly IDocumentTrackingService _documentTrackingService;
 
     /// <summary>
@@ -54,16 +48,12 @@ internal sealed class SolutionChecksumUpdater
     private const string SynchronizeTextChangesStatusSucceededKeyName = nameof(SolutionChecksumUpdater) + "." + SynchronizeTextChangesStatusSucceededMetricName;
     private const string SynchronizeTextChangesStatusFailedKeyName = nameof(SolutionChecksumUpdater) + "." + SynchronizeTextChangesStatusFailedMetricName;
 
-    private bool _isSynchronizeWorkspacePaused;
-
     public SolutionChecksumUpdater(
         Workspace workspace,
         IAsynchronousOperationListenerProvider listenerProvider,
         CancellationToken shutdownToken)
     {
         var listener = listenerProvider.GetListener(FeatureAttribute.SolutionChecksumUpdater);
-
-        _globalOperationService = workspace.Services.SolutionServices.ExportProvider.GetExports<IGlobalOperationNotificationService>().FirstOrDefault()?.Value;
 
         _workspace = workspace;
         _documentTrackingService = workspace.Services.GetRequiredService<IDocumentTrackingService>();
@@ -87,12 +77,6 @@ internal sealed class SolutionChecksumUpdater
         _workspaceChangedImmediateDisposer = _workspace.RegisterWorkspaceChangedImmediateHandler(OnWorkspaceChangedImmediate);
         _documentTrackingService.ActiveDocumentChanged += OnActiveDocumentChanged;
 
-        if (_globalOperationService != null)
-        {
-            _globalOperationService.Started += OnGlobalOperationStarted;
-            _globalOperationService.Stopped += OnGlobalOperationStopped;
-        }
-
         // Enqueue the work to sync the initial data over.
         _synchronizeActiveDocumentQueue.AddWork();
         _synchronizeWorkspaceQueue.AddWork();
@@ -101,53 +85,21 @@ internal sealed class SolutionChecksumUpdater
     public void Shutdown()
     {
         // Try to stop any work that is in progress.
-        PauseSynchronizingPrimaryWorkspace();
+        lock (_gate)
+        {
+            _synchronizeWorkspaceQueue.CancelExistingWork();
+        }
 
         _documentTrackingService.ActiveDocumentChanged -= OnActiveDocumentChanged;
         _workspaceChangedDisposer.Dispose();
         _workspaceChangedImmediateDisposer.Dispose();
-
-        if (_globalOperationService != null)
-        {
-            _globalOperationService.Started -= OnGlobalOperationStarted;
-            _globalOperationService.Stopped -= OnGlobalOperationStopped;
-        }
-    }
-
-    private void OnGlobalOperationStarted(object? sender, EventArgs e)
-        => PauseSynchronizingPrimaryWorkspace();
-
-    private void OnGlobalOperationStopped(object? sender, EventArgs e)
-        => ResumeSynchronizingPrimaryWorkspace();
-
-    private void PauseSynchronizingPrimaryWorkspace()
-    {
-        // An expensive global operation started (like a build).  Pause ourselves and cancel any outstanding work in
-        // progress to synchronize the solution.
-        lock (_gate)
-        {
-            _synchronizeWorkspaceQueue.CancelExistingWork();
-            _isSynchronizeWorkspacePaused = true;
-        }
-    }
-
-    private void ResumeSynchronizingPrimaryWorkspace()
-    {
-        lock (_gate)
-        {
-            _isSynchronizeWorkspacePaused = false;
-            _synchronizeWorkspaceQueue.AddWork();
-        }
     }
 
     private void OnWorkspaceChanged(WorkspaceChangeEventArgs _)
     {
-        // Check if we're currently paused.  If so ignore this notification.  We don't want to any work in response
-        // to whatever the workspace is doing.
         lock (_gate)
         {
-            if (!_isSynchronizeWorkspacePaused)
-                _synchronizeWorkspaceQueue.AddWork();
+            _synchronizeWorkspaceQueue.AddWork();
         }
     }
